Position -> bool isLong; uint112 ioShares; uint88 cost; uint88 debt; uint256 entryPrice;

+++ contracts/OverlayV1Market.sol
@@ -146,7 +146,7 @@ contract OverlayV1Market {
             msg.sender,
             positionId_,
             Position.Info({
-                oiShares: uint88(oi),
+                oiShares: uint112(oi),
                 cost: uint88(collateral),
                 debt: uint88(oi - collateral),
                 isLong: isLong,
